Migrate prefabNames,PrefabTiers,HullMaterial to unity shared lib
