+++ src/ValheimRAFT.Unity/Assets/ValheimVehicles/SharedScripts/Enums/HullMaterial.cs
@@ -1,0 +1,11 @@
+﻿namespace ValheimVehicles.SharedScripts.Enums
+{
+  public abstract class HullMaterial
+  {
+    public const string CoreWood = "core_wood";
+    public const string ElderWood = "elder_wood";
+    public const string Wood = "wood";
+    public const string Iron = "iron";
+    public const string YggdrasilWood = "yggdrasil_wood";
+  }
+}
