+++ src/ValheimVehicles/ValheimVehicles.Prefabs/ShipHulls.cs
@@ -1,4 +1,5 @@
-namespace ValheimVehicles.Prefabs;
+using ValheimVehicles.SharedScripts.Enums;
+namespace ValheimVehicles.SharedScripts.Prefabs;
 
 public abstract class ShipHulls
 {
@@ -7,16 +8,7 @@ public abstract class ShipHulls
     Horizontal = 0,
     Vertical = 90,
     FortyFive = 45,
-    TwentySix = 26,
-  }
-
-  public abstract class HullMaterial
-  {
-    public const string CoreWood = "core_wood";
-    public const string ElderWood = "elder_wood";
-    public const string Wood = "wood";
-    public const string Iron = "iron";
-    public const string YggdrasilWood = "yggdrasil_wood";
+    TwentySix = 26
   }
 
   /// <summary>
@@ -24,12 +16,14 @@ public abstract class ShipHulls
   /// </summary>
   /// <param name="prefabName"></param>
   /// <returns></returns>
-  public static bool GetExcludedBoundsPrefabCollider(string prefabName) =>
-    prefabName.StartsWith(PrefabNames.HullProw) || prefabName.StartsWith(PrefabNames.HullRibCorner);
+  public static bool GetExcludedBoundsPrefabCollider(string prefabName)
+  {
+    return prefabName.StartsWith(PrefabNames.HullProw) || prefabName.StartsWith(PrefabNames.HullRibCorner);
+  }
 
   public static string GetHullMaterialDescription(string materialVariant)
   {
-    return materialVariant == ShipHulls.HullMaterial.Wood
+    return materialVariant == HullMaterial.Wood
       ? "$valheim_vehicles_material_wood_desc"
       : "$valheim_vehicles_material_iron_yggdrasil_desc";
   }
@@ -41,7 +35,7 @@ public abstract class ShipHulls
       // wood health support is 400 for a normal plank but these are extra thick so +200
       // iron reinforced wood is 2000 health but these are rustic plats but can be made larger and are also slightly weaker since they incorporate bronze out layer
       case HullMaterial.Iron:
-        PrefabRegistryHelpers.SetWearNTearSupport(wnt, WearNTear.MaterialType.Iron);
+        // PrefabRegistryHelpers.SetWearNTearSupport(wnt, WearNTear.MaterialType.Iron);
         wnt.m_health = 500 * pieces;
         break;
       case HullMaterial.Wood:
