+++ src/ValheimVehicles/ValheimVehicles.csproj
@@ -29,7 +29,9 @@
     <ItemGroup Label="ValheimVehicle.SharedScripts Unity testable scripts">
         <!--  Allows for both sharing and testing these scripts within unity. Best of both worlds.      -->
         <Compile Include="..\ValheimRAFT.Unity\Assets\ValheimVehicles\SharedScripts\**\*.cs"/>
-        <Folder Include="..\ValheimRAFT.Unity\Assets\ValheimVehicles\SharedScripts\"/>
+        <Compile Update="..\ValheimRAFT.Unity\Assets\ValheimVehicles\SharedScripts\PrefabNames.cs">
+            <Link>SharedScripts\PrefabNames.cs</Link>
+        </Compile>
     </ItemGroup>
     <ItemGroup Label="Internal ValheimVehicles scripts">
         <!-- Includes all namespace folders -->
